Cleanup and reorganise components.
