+++ app/js/components/logic/README.md
@@ -1,0 +1,3 @@
+# Logic Components
+
+All components that are statefull, i.e. they have `state` to track are stored here.
