+++ app/js/components/view/ipfs-logo.js
@@ -9,11 +9,13 @@ const styles = {
   }
 }
 
+const logoPath = require('../../../../node_modules/ipfs-logo/ipfs-logo-wider-strokes-white.png')
+
 export default function IPFSLogo () {
   return (
     <div style={styles.wrapper}>
       <image
-        src={require('../../../node_modules/ipfs-logo/ipfs-logo-wider-strokes-white.png')}
+        src={logoPath}
         width={16}
         height={16}
         style={styles.image}
