+++ app/js/menubar.js
@@ -1,7 +1,7 @@
 import {render} from 'react-dom'
 import React from 'react'
 
-import Menu from './components/menu'
+import Menu from './screens/menu'
 
 document.addEventListener('DOMContentLoaded', () => {
   render(<Menu />, document.getElementById('menubar'))
