+++ app/js/screens/README.md
@@ -1,0 +1,5 @@
+# Screens
+
+A screen is the everything visible at a given point in time inside a single window.
+These are implemented as react components, but differ in so far as they are an aggregation
+of multiple small components to one larger, complete picture.
