+++ app/js/screens/setup/intro.js
@@ -1,11 +1,11 @@
-import React, {PropTypes} from 'react'
+import React, {Component, PropTypes} from 'react'
 import Radium from 'radium'
 
-import Icon from '../../icon'
-import Button from '../../button'
+import Icon from '../../components/view/icon'
+import Button from '../../components/view/button'
 
 @Radium
-export default class Intro extends React.Component {
+export default class Intro extends Component {
 
   static propTypes = {
     onInstallClick: PropTypes.func,
@@ -23,7 +23,7 @@ export default class Intro extends React.Component {
         display: 'flex',
         width: '100%',
         height: '100%',
-        backgroundImage: `url(${require('../../../../img/jellyfish-large.png')})`,
+        backgroundImage: `url(${require('../../../img/jellyfish-large.png')})`,
         backgroundSize: '100%',
         backgroundPosition: '0 0',
         color: '#FFFFFF',
