+++ app/js/welcome.js
@@ -1,7 +1,7 @@
 import {render} from 'react-dom'
 import React from 'react'
 
-import Setup from './components/setup'
+import Setup from './screens/setup'
 
 document.addEventListener('DOMContentLoaded', () => {
   render(<Setup />, document.getElementById('welcome'))
