--- app/views/help.html
@@ -1,14 +1,0 @@
-<html>
-  <head>
-    <title>IPFS - Error found, apologies for the inconvience</title>
-    <meta charset="utf-8"/>
-    <script src="help.js"></script>
-  </head>
-  <body>
-    <div id="error_row">
-      Error found, apologies for the inconvience
-    </div>
-    <div id="error">
-    </div>
-  </body>
-</html>
